refactor ParallelFlowExecutorService
