+++ src/main/java/executor/service/service/ParallelFlowExecutorService.java
@@ -1,23 +1,7 @@
 package executor.service.service;
 
-/**
- * Start ExecutionService in parallel multi-threaded mode.
- *
- * @author Oleksandr Tuleninov
- * @version 01
- */
-public interface ParallelFlowExecutorService {
+import java.util.concurrent.ExecutorService;
 
-    /**
-     * Start ScenarioSourceListener, ProxySourcesClient, ExecutionService
-     * in parallel multi-threaded mode.
-     */
-    void execute();
-
-    /**
-     * Initiates an orderly shutdown in which previously submitted tasks are executed,
-     * but no new tasks will be accepted.
-     */
-    void shutdown();
+public interface ParallelFlowExecutorService extends ExecutorService {
 
 }
